pyproject.toml
[project]
name = "auto-scan"
version = "0.1.0"
description = "Scan documents from any network scanner and auto-classify with AI"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.28,<1",
    "zeroconf>=0.148,<1",
    "anthropic>=0.94,<1",
    "Pillow>=11,<13",
    "img2pdf>=0.6,<1",
    "pikepdf>=9,<11",
    "python-dotenv>=1.2,<2",
    "flask>=3.1,<4",
    "pytesseract>=0.3,<1",
]

[project.scripts]
auto-scan = "auto_scan.cli:main"
auto-scan-gui = "auto_scan.gui:main"

[project.optional-dependencies]
test = [
    "pytest>=8,<9",
    "pytest-cov>=6,<7",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]
